pyproject.toml
[tool.poetry]
name = "datakube"
version = "0.1.0"
description = "Data analysis for Kubernetes data metrics"
authors = ["David Morrison <drmorr@appliedcomputing.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
pyarrow = "^16.0.0"
arrow = "^1.3.0"
kubernetes-client = "^0.1.8"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
pytest = "^8.1.1"
ipykernel = "^6.29.4"
pandas-stubs = "^2.2.1.240316"
ruff = "^0.4.2"
kubernetes-stubs = "^22.6.0.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
preview = true
select = ["E", "F", "I", "W", "PL"]
ignore = ["PLR2004"]

[tool.ruff.lint.isort]
force-single-line = true
